docs: move Storybook to docs application
Given we could have in the future other packages, e.g.
a version of DT UI for Web components, it doesn't make sense
to have the generic documentation on dt-ui-react package,
and makes more sense to have in a generic place on the
repository, in this case in a dedicated application for the
documentation.

We also decided to keep the component's stories alongside
with the components, to make the developer flow easier.
That's why we are getting the stories from dt-ui-react
package.