fix: debounce call on pagination
The debounced function was being called in a useEffect that triggered on
every input value change. Instead, it can be called directly inside the
handleChange function.
This will allow passing an onChange with changes on the query params,
using Next Router. Otherwise, it would cause an infinite loop since we
needed to put the router in a dependency array.