chore: update concurrency logic
In storybook workflow, we can cancel the current flow if we are still in
the build-storybook job and if the trigger matches a flow that is
currently running.
In the main workflow, we should add a fallback to the group name since
`github.event.head_commit.id` will only work on pushes and not on
pull_requests.