chore: update publish trigger
Github Actions has a condition to only trigger an event to a max of
three tags pushed at the same time. Since we have a lot of independent
packages in DT-UI, we need to use a different approach to trigger the
publish step. In this PR, we use the trigger of pushing to main branch
and check if the commit message includes the release message.