chore: add cleanup stage for gh-pages
Sometimes, the step for deleting PR previews on gh-pages branch may fail
or we might close a PR without merging it, ending up with unnecessary
code in the gh-pages branch. This code will take up space that might be
needed for future PRs. This cleanup stage will run every monday to clean
up unused folders by comparing the open PRs with what is currently in
the gh-pages branch.