feat: add id to base props
As a developer, I want to be able to pass an ID to the typography
component. To enable this functionality for other components as well, we
aim to add the ID property to the base props so that it can be forwarded
to those components.

There are several cases in which the ID is used. However, I want to use
it because of accessibility. The aria-labelledby and aria-describedby
attributes also use IDs to link descriptive text to elements, improving
accessibility.